global: legacy code removal
* INCOMPATIBLE Removes legacy sample workflows, tasks and
  definitions. (closes #14)

* Updates tests to be compatible with these changes.
