+++ invenio_workflows/definitions.py
@@ -19,10 +19,6 @@
 
 """Contains basic workflow types for use in workflow definitions."""
 
-import collections
-
-from six import string_types
-
 
 class WorkflowMissing(object):
 
@@ -63,129 +59,5 @@ class WorkflowBase(object):
 
     @staticmethod
     def get_sort_data(obj, **kwargs):
-        """Return a dictionary of key values useful for sorting in Holding Pen."""
+        """Return a dictionary useful for sorting in Holding Pen."""
         return {}
-
-
-class RecordWorkflow(WorkflowBase):
-
-    """Workflow to be used where BibWorkflowObject is a Record instance."""
-
-    workflow = []
-
-    @staticmethod
-    def get_title(bwo):
-        """Get the title."""
-        field_map = {"title": "title"}
-        try:
-            record = bwo.data
-        except AttributeError:
-            record = bwo.get_data()
-        extracted_titles = []
-        if hasattr(record, "get") and "title" in record:
-            if isinstance(record["title"], str):
-                extracted_titles = [record["title"]]
-            else:
-                extracted_titles.append(record["title"][field_map["title"]])
-        return ", ".join(extracted_titles) or "No title found"
-
-    @staticmethod
-    def get_description(bwo):
-        """Get the description (identifiers and categories) from the object data."""
-        from flask import render_template
-
-        try:
-            record = bwo.data
-        except AttributeError:
-            record = bwo.get_data()
-        # Get identifiers
-        final_identifiers = []
-        categories = []
-
-        if hasattr(record, "get"):
-            doi = record.get("doi")
-            if doi:
-                final_identifiers.append(doi)
-
-            system_no = record.get("system_number_external", {}).get("value")
-            if system_no:
-                final_identifiers.append(system_no)
-
-            if 'subject' in record:
-                lookup = ["subject", "term"]
-            elif "subject_term" in record:
-                lookup = ["subject_term", "term"]
-            else:
-                lookup = None
-            if lookup:
-                primary, secondary = lookup
-                category_list = record.get(primary, [])
-                if isinstance(category_list, dict):
-                    category_list = [category_list]
-                categories = [subject[secondary] for subject in category_list]
-
-        return render_template('workflows/styles/harvesting_record.html',
-                               categories=categories,
-                               identifiers=final_identifiers)
-
-    @staticmethod
-    def formatter(bwo, **kwargs):
-        """Nicely format the record."""
-        from pprint import pformat
-        from invenio_records.api import Record
-
-        try:
-            data = bwo.data
-        except AttributeError:
-            data = bwo.get_data()
-
-        if not data:
-            return ''
-
-        formatter = kwargs.get("formatter", None)
-        of = kwargs.get("of", None)
-        if formatter:
-            # A separate formatter is supplied
-            return formatter(data)
-
-        if isinstance(data, collections.Mapping):
-            # Dicts are cool on its own, but maybe its SmartJson (record)
-            try:
-                data = Record(data.dumps()).legacy_export_as_marc()
-            except (TypeError, KeyError):
-                pass
-
-        if isinstance(data, string_types):
-            # We can try formatter!
-            # If already XML, format_record does not like it.
-            if of and of != 'xm':
-                try:
-                    from invenio_formatter import format_record
-                    formatted_data = format_record(
-                        recID=None,
-                        of=of,
-                        xml_record=data
-                    )
-                except TypeError:
-                    # Wrong kind of type
-                    pass
-            else:
-                # So, XML then
-                from xml.dom.minidom import parseString
-
-                try:
-                    unpretty_data = parseString(data)
-                    formatted_data = unpretty_data.toprettyxml()
-                except TypeError:
-                    # Probably not proper XML string then
-                    return "Data cannot be parsed: %s" % (data,)
-                except Exception:
-                    # Just return raw string
-                    pass
-
-        if not formatted_data:
-            formatted_data = data
-
-        if isinstance(formatted_data, dict):
-            formatted_data = pformat(formatted_data)
-        return formatted_data
