+++ invenio_workflows/workflows/sample_workflow.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2012, 2013, 2014 CERN.
+# Copyright (C) 2012, 2013, 2014, 2015 CERN.
 #
 # Invenio is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License as
@@ -20,9 +20,8 @@
 """Showcase a sample workflow definition."""
 
 from ..definitions import WorkflowBase
-from ..tasks.sample_tasks import add_data
 from ..tasks.logic_tasks import execute_if
-from ..tasks.marcxml_tasks import approve_record
+from ..tasks.sample_tasks import add_data, approve_record
 
 
 class sample_workflow(WorkflowBase):
